#9-refactor: 세부 구현에 해당하는 메서드의 접근 제어자를 private로 바꾸었다..
